+++ vini_app/frmExportMouvementArticle.vb
@@ -155,14 +155,18 @@ Public Class frmExportMouvementArticle
         '
         'WebBrowser1
         '
+        Me.WebBrowser1.AllowWebBrowserDrop = False
         Me.WebBrowser1.Anchor = CType((((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Bottom) _
             Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
+        Me.WebBrowser1.IsWebBrowserContextMenuEnabled = False
         Me.WebBrowser1.Location = New System.Drawing.Point(13, 136)
         Me.WebBrowser1.MinimumSize = New System.Drawing.Size(20, 20)
         Me.WebBrowser1.Name = "WebBrowser1"
+        Me.WebBrowser1.ScriptErrorsSuppressed = True
         Me.WebBrowser1.Size = New System.Drawing.Size(933, 484)
         Me.WebBrowser1.TabIndex = 15
+        Me.WebBrowser1.WebBrowserShortcutsEnabled = False
         '
         'frmExportMouvementArticle
         '
@@ -192,6 +196,7 @@ Public Class frmExportMouvementArticle
 
     Private Sub frmMouvementArticle_Load(ByVal sender As Object, ByVal e As System.EventArgs) Handles MyBase.Load
         EnableControls(True)
+        Me.WebBrowser1.Visible = False
     End Sub
 
     Private Sub Label3_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Label3.Click
@@ -199,39 +204,48 @@ Public Class frmExportMouvementArticle
     End Sub
 
     Private Sub cbExporter_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles cbExporter.Click
-        setcursorWait()
         ExporterMvtArticles()
-        restoreCursor()
     End Sub
 
     Private Sub ExporterMvtArticles()
         'On Prend les fournisseurs indiqué comme Espace fournisseur
+        lblProgress.Text = "Lecture des fournisseurs"
         m_ListFRN = Fournisseur.getListe().Where(Function(o) o.EspFrn).ToList()
+        If tbCodeFourn.Text <> "" Then
+            m_ListFRN = m_ListFRN.Where(Function(o) o.code = tbCodeFourn.Text).ToList()
+        End If
         'm_ListFRN = Fournisseur.getListe(tbCodeFourn.Text)
 
-        m_pgBar.Maximum = m_ListFRN.Count() + 1
-        m_pgBar.Minimum = 0
-        m_pgBar.Value = 0
+        If m_ListFRN.Count > 0 Then
+            m_pgBar.Maximum = m_ListFRN.Count() + 1
+            m_pgBar.Minimum = 0
+            m_pgBar.Value = 0
 
-        m_strFolder = My.Settings.Tmp & "/ExportMvtArticle" & DateTime.Now.ToString("yyyyMMdd")
-        If Not System.IO.Directory.Exists(m_strFolder) Then
-            System.IO.Directory.CreateDirectory(m_strFolder)
-        End If
-        If Directory.EnumerateFiles(m_strFolder).Count > 0 Then
-            If MsgBox("Votre répertoire " & m_strFolder & " n'est pas vide, souhaitez-vous supprimer les fichiers existants avant l'export ?", MsgBoxStyle.YesNo) = MsgBoxResult.Yes Then
-                For Each strFile As String In Directory.EnumerateFiles(m_strFolder)
-                    File.Delete(strFile)
-                Next
+            m_strFolder = My.Settings.Tmp & "/ExportMvtArticle" & DateTime.Now.ToString("yyyyMMdd")
+            If Not System.IO.Directory.Exists(m_strFolder) Then
+                System.IO.Directory.CreateDirectory(m_strFolder)
             End If
+            If Directory.EnumerateFiles(m_strFolder).Count > 0 Then
+                If MsgBox("Votre répertoire " & m_strFolder & " n'est pas vide, souhaitez-vous supprimer les fichiers existants avant l'export ?", MsgBoxStyle.YesNo) = MsgBoxResult.Yes Then
+                    For Each strFile As String In Directory.EnumerateFiles(m_strFolder)
+                        File.Delete(strFile)
+                    Next
+                End If
+            End If
+
+            m_strFileCSV = m_strFolder & "/mvt" & dtDFin.Value.Year & dtDFin.Value.Month & dtDFin.Value.Day & ".csv"
+
+
+            Me.Cursor = Cursors.WaitCursor
+            Me.UseWaitCursor = True
+            lblProgress.Text = "Traitement des fournisseurs"
+
+            BackgroundWorker1.RunWorkerAsync()
+            cbExporter.Enabled = False
+        Else
+            lblProgress.Text = ""
+            MsgBox("Aucun fournisseur à traiter")
         End If
-
-        m_strFileCSV = m_strFolder & "/mvt" & dtDFin.Value.Year & dtDFin.Value.Month & dtDFin.Value.Day & ".csv"
-
-
-
-        BackgroundWorker1.RunWorkerAsync()
-        cbExporter.Enabled = False
-        Me.Cursor = Cursors.WaitCursor
     End Sub
 
     Private Sub BackgroundWorker1_DoWork(sender As Object, e As System.ComponentModel.DoWorkEventArgs) Handles BackgroundWorker1.DoWork
@@ -273,7 +287,11 @@ Public Class frmExportMouvementArticle
     End Sub
 
     Private Sub BackgroundWorker1_RunWorkerCompleted(sender As Object, e As System.ComponentModel.RunWorkerCompletedEventArgs) Handles BackgroundWorker1.RunWorkerCompleted
+        Me.WebBrowser1.ScriptErrorsSuppressed = True
+        Me.WebBrowser1.Visible = True
         Me.WebBrowser1.Navigate(New Uri(Param.getConstante("CST_FTPVNC_URL")))
+        m_pgBar.Value = m_pgBar.Value + 1
+        lblProgress.Text = m_pgBar.Value & "/" & m_pgBar.Maximum
     End Sub
 
     Private Sub btn_annuler_Click(sender As Object, e As EventArgs) Handles btn_annuler.Click
@@ -282,6 +300,7 @@ Public Class frmExportMouvementArticle
 
     Private Sub WebBrowser1_DocumentCompleted(sender As Object, e As WebBrowserDocumentCompletedEventArgs) Handles WebBrowser1.DocumentCompleted
         Me.Cursor = Cursors.Default
+        Me.UseWaitCursor = False
         MsgBox("export terminé")
         cbExporter.Enabled = True
     End Sub
